backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
python-dotenv
sqlalchemy
